+++ amberdata/src/endpoint/price.ts
@@ -7,9 +7,15 @@ const customError = (data: any) => {
   return Object.keys(data.payload).length === 0
 }
 
+const addressMapping: { [symbol: string]: string } = {
+  DIGG: '0x798d1be841a82a273720ce31c822c61a67a601c3',
+  WBTC: '0x2260fac5e5542a773aa44fbcfedf7c193bc2c599',
+}
+
 const customParams = {
   base: ['base', 'from', 'coin'],
   quote: ['quote', 'to', 'market'],
+  includes: false,
 }
 
 export const execute: ExecuteWithConfig<Config> = async (input, config) => {
@@ -19,10 +25,22 @@ export const execute: ExecuteWithConfig<Config> = async (input, config) => {
 
   const coin = validator.validated.data.base
   const market = validator.validated.data.quote
-  const url = `/api/v2/market/spot/prices/pairs/${coin.toLowerCase()}_${market.toLowerCase()}/latest`
+  const includes = validator.validated.data.includes || []
 
-  const params = {
+  let url = `/api/v2/market/spot/prices/pairs/${coin.toLowerCase()}_${market.toLowerCase()}/latest`
+  let params: { [key: string]: any } = {
     includeCrossRates: true,
+  }
+
+  if (
+    includes.length > 0 &&
+    includes[0].toLowerCase() === 'wbtc' &&
+    coin.toLowerCase() === 'digg'
+  ) {
+    const fromAddress = addressMapping[coin.toUpperCase()]
+    const toAddress = addressMapping[includes[0].toUpperCase()]
+    url = `/api/v2/market/defi/prices/pairs/bases/${fromAddress}/quotes/${toAddress}/latest`
+    params = {}
   }
 
   const reqConfig = { ...config.api, params, url }