+++ therundown/src/config.ts
@@ -1,0 +1,18 @@
+import { Requester } from '@chainlink/external-adapter'
+import { Config } from '@chainlink/types'
+
+export const DEFAULT_ENDPOINT = 'total-score'
+export const DEFAULT_BASE_URL = 'https://therundown-therundown-v1.p.rapidapi.com/'
+
+export const makeConfig = (prefix?: string): Config => {
+  const config = Requester.getDefaultConfig(prefix, true)
+  config.api = {
+    ...config.api,
+    baseURL: config.api.baseURL || DEFAULT_BASE_URL,
+    headers: {
+      ...config.api.headers,
+      'x-rapidapi-host': 'therundown-therundown-v1.p.rapidapi.com',
+    },
+  }
+  return config
+}