+++ trueusd/src/config.ts
@@ -1,14 +1,11 @@
 import { Requester } from '@chainlink/external-adapter'
 import { Config } from '@chainlink/types'
 
-export const DEFAULT_ENDPOINT = 'matches'
-export const DEFAULT_BASE_URL = 'http://35.195.237.123:8000/'
+export const DEFAULT_ENDPOINT = 'trueusd'
+export const DEFAULT_BASE_URL = 'https://core-api.real-time-attest.trustexplorer.io'
 
 export const makeConfig = (prefix?: string): Config => {
   const config = Requester.getDefaultConfig(prefix)
-  config.api = {
-    ...config.api,
-    baseURL: config.api.baseUrl || DEFAULT_BASE_URL,
-  }
+  config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
   return config
 }