+++ CHANGELOG.md
@@ -10,10 +10,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - New adapters:
   - `agoric` to push results to the Agoric blockchain
+  - `therundown` to get sports score data from TheRundown
+  - `tradingeconomics-stream` to get stream data from TradingEconomics
 
 ### Changed
 
-- Updated support for `DIGG/BTC` in Kaiko
+- Updated support for `DIGG/BTC` in Kaiko and Amberdata
+- Updated base URL for GeoDB
+- CMC now uses preset IDs instead of preset slugs
 
 ## [0.2.0-rc.1] - 2021-2-4